+++ SourceCode/CrySoundSystem/MusicPatternInstance.h
@@ -27,7 +27,7 @@ public:
 	CMusicPattern* GetPattern() { return m_pPattern; }
 	//! Seek to beginning of pattern (if nDelay is set it will wait nDelay-samples before starting playback).
 	bool Seek0(int nDelay=0);
-	bool GetPCMData(signed long *pDataOut, int nSamples, bool bLoop=true);
+	bool GetPCMData(signed int *pDataOut, int nSamples, bool bLoop=true);
 	int GetSamplesToNextFadePoint();
 	int GetSamplesToLastFadePoint();
 	int GetSamplesToEnd();
